box-office: using babel instead of esm
