+++ workspaces/box-office/.gitignore
@@ -1,3 +1,4 @@
 # telepath.config
 idbox-firebase.json
 Services.json
+lib
