+++ workspaces/box-office/source/entry-point/EntryPoint.js
