+++ workspaces/box-office/source/entry-point/index.js
