+++ workspaces/box-office/source/services/ServiceRegistrationService.js
