+++ workspaces/box-office/source/services/ServiceRegistry.js
